PDI-1278 - Refactor dialog sizing/positioning code
Use primary monitor client area instead of bounds.  This takes OS chrome into account.
If the size passed in is too large for the primary monitor client area, it is shrunk to fit. If the positioning leaves part of the dialog outside the client area, it is centered instead.

Note that currently, many of the defaults in org.pentaho.di.ui.core/default.properties have crazy values. This causes the failsafe code in here to fire a lot more than is really necessary.


git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@8445 5fb7f6ec-07c1-534a-b4ca-9155e429e800