PDI-12766 - NPEs in Spoon UI when running transformation with partitioning and clustering set up
This is fix to clustered/partitioned transformation execution
configuration.
This NPE is specific when transformation is desigend to run on a cluster
and no repositories or shared xml files is available.
For this case we used to lose partition scheam when sending transformation
in form of xml.
As said above if repository is available this info will be read locally
and NPE will be avoided.

Also repair originally commented check when step running in multiple
copies in cluster connected to non-clustered step which is running not
clustered. We used to have NPE for this case so now error message will be
shown suggest to customer to use dummy step if necessary.

Repaired log messages.