[BACKLOG-186] - Need to provide user with a decent message when the repository is not availabe
since the UI is formed so that it's impossible to perform centralized error handling
it was decided to find as many places as it's only possible and implement the check

the check checks that Cause-chain contain KettleRepositoryLostException.
if it happens standard Spoon ErrorDialog is displayed, then current form is closed
and at last a command is issued to spoon to close the current repository connection.

Conflicts:
	ui/src/org/pentaho/di/ui/repository/repositoryexplorer/controllers/BrowseController.java