[SP-2786] Backport of PDI-15120 - Export Linked Resources to XML fails if Text File Input contains empty row (6.1 Suite)
- Adding null-checks before obtaining File Object from filename, that can be null.

Summary:

After this fix behavior is the following:
- When exporting trans with empty filenames, they are kept empty and not converted to related path

More details:

Situation is the following here:
In transformation's xml fileNames with empty strings are stored as <fileName/> and extracted as nulls when transformation is loaded.
First I wanted to store and load them as empty strings. But it was a bad idea - it leads us to some problems that are hard to overcome with such storing. Here the pitfall of this approach: when exporting trans with empty (non-null) strings in file name - they are converted into related path (root of kettle in this case). And in 99 % cases it's not what user expects.

So what we want here is exporting empty strings as they are, without converting them to related path.
We should also do this with empty strings for keeping consistency. Try running trans with textFileInput step with empty file name. It won't be treated as related path. But when exporting if with empty string it becomes related, that is not expected.
(cherry picked from commit 25aaeb6)