[PDI-4048] Text file input cannot properly handle a fixed-length-of-bytes file - Adding 'length' option so that a user can specify which length ('Characters' or 'Bytes').
Currently, a file containing multi-byte characters can be properly parsed only when the length is character-based.
This issue occurs when a file contains multi-byte characters and each field has a fixed-length of bytes.
This is not an unusual case since columns of a database table may be defined by bytes instead of characters and so are the exported fixed-length files.