[BACKLOG-19327] - Transformation runs again instead of getting data from cache.
The transmeta XML is being used to generate the chache key for the dataservices. DET state (attributes groups) is being stored in the TransMeta which produces changes in the XML, invalidating the data services cache. It was added an overload getXML method to be able to omit the attributes groups.