[PDI-18619] Memory leak due to failing to clean up EmbeddedMetastore
Within a single job or trans, multiple other jobs/trans can be launched in
parallel. It's also possible for more than one of those jobs/trans to share
a jobmeta or transmeta. Because of this, the former logic for disposing
and creating the metastorelocator key could result in orphaned keys,
since multiple threads may create new ones without disposing the previous.

This change (combined with pentaho/pentaho-osgi-bundles#364)
stores only a single embeddedmetastore for each meta, and associates a single
key. Disposal verifies that the trans or job doing the disposing is the "top level"
job or trans holding that key, making sure it will be cleaned up when no other job
or trans needs it.

https://jira.pentaho.com/browse/PDI-18619