[PDI-18719] Global LogChannels do not pick up DefaultLogLevel changes
The global log channels in PDI are initialized on startup with the
default log level of BASIC.

The default log level (loaded from .spoonrc|LogLevel) is read after these
fields are initialized, so those logchannels are not updated.  When the
default log level is changed we should update the global log channels.

https://jira.pentaho.com/browse/PDI-18719