[BACKLOG-44137] - improve db loading from Repository
- with projects, the runtime database for a step can change based on which
project is used. This means that a step might be saved referring to a
DatabaseMeta by ObjectId, but at load time, the DatabaseMeta with that
ObjectId is not present in the execution-time list. To address this, we
now load the original DatabaseMeta by ObjectId, and then use that
DatabaseMeta's name to find the correct one in the execution-time list.