[BACKLOG-44647] - propagate changes from RepositoryBowl vfs connections
DefaultBowl and RepositoryBowl each have a ConnectionManager instance that
caches connections. The RepositoryBowl is used to make edits.

At runtime, code that has been updated to support Bowls will use the
RepositoryBowl to access VFS connections, but code that has not been updated
will use legacy APIs that use DefaultBowl. DefaultBowl will still find the
connections from the repository metastore thanks to using the metastore
locator, but it caches the results. That means that if an edit is made in
RepositoryBowl, it will not be reflected in code that uses DefaultBowl.

This change adds cache invalidation to DefaultBowl ConnectionManager whenever
edits are made to the RepositoryBowl ConnectionManager.