[BACKLOG-44486] - Thread safety for SharedObjectsIO, Node
org.w3c.dom.Node is completely non-thread-safe, even for reads. This means we need
to lock around all access to Node objects, which means all the calls into SharedObjectsIO
that pass Nodes around.

This locking also has to take DelegatingSharedObjectsIO into account, so the API is
at the SharedObjectsIO level, and DelegatingSharedObjectsIO locks each of the supplying
instances.

This mostly only needs to be exposed to the Manager classes that use SharedObjectsIO,
but AbstractMeta, JobMeta, and TransMeta need careful handling of the locks around
their clone() methods. And the tests, of course.